+++ packages/edit-site/src/components/sidebar-edit-mode/template-panel/template-actions.js
@@ -1,12 +1,10 @@
 /**
  * WordPress dependencies
  */
-import { useDispatch, useSelect } from '@wordpress/data';
+import { useDispatch } from '@wordpress/data';
 import { __ } from '@wordpress/i18n';
 import { DropdownMenu, MenuGroup, MenuItem } from '@wordpress/components';
 import { moreVertical } from '@wordpress/icons';
-import { parse } from '@wordpress/blocks';
-import { store as coreStore } from '@wordpress/core-data';
 
 /**
  * Internal dependencies
@@ -14,53 +12,16 @@ import { store as coreStore } from '@wordpress/core-data';
 import { store as editSiteStore } from '../../../store';
 import isTemplateRevertable from '../../../utils/is-template-revertable';
 import ReplaceTemplateButton from './replace-template-button';
-import { unlock } from '../../../lock-unlock';
-import { PATTERN_CORE_SOURCES, PATTERN_TYPES } from '../../../utils/constants';
-
+import { useAvailablePatterns } from '../page-panels/hooks';
 export default function Actions( { template } ) {
-	// This is duplicated.
-	const filterOutDuplicatesByName = ( currentItem, index, items ) =>
-		index === items.findIndex( ( item ) => currentItem.name === item.name );
-
-	// Should we also get templates?
-	const availableTemplates = useSelect( ( select ) => {
-		const { getSettings } = unlock( select( editSiteStore ) );
-		const settings = getSettings();
-		const blockPatterns =
-			settings.__experimentalAdditionalBlockPatterns ??
-			settings.__experimentalBlockPatterns;
-
-		const restBlockPatterns = select( coreStore ).getBlockPatterns();
-
-		const patterns = [
-			...( blockPatterns || [] ),
-			...( restBlockPatterns || [] ),
-		]
-			.filter(
-				( pattern ) => ! PATTERN_CORE_SOURCES.includes( pattern.source )
-			)
-			.filter( filterOutDuplicatesByName )
-			// Filter only the patterns that are compatible with the current template.
-			.filter( ( pattern ) =>
-				pattern.templateTypes?.includes( template.slug )
-			)
-			.map( ( pattern ) => ( {
-				...pattern,
-				keywords: pattern.keywords || [],
-				type: PATTERN_TYPES.theme,
-				blocks: parse( pattern.content, {
-					__unstableSkipMigrationLogs: true,
-				} ),
-			} ) );
-
-		return patterns.map( ( pattern ) => {
-			return {
-				name: pattern.name,
-				blocks: pattern.blocks,
-				title: pattern.title,
-				content: pattern.content,
-			};
-		} );
+	const availablePatterns = useAvailablePatterns( template );
+	const availableTemplates = availablePatterns.map( ( pattern ) => {
+		return {
+			name: pattern.name,
+			blocks: pattern.blocks,
+			title: pattern.title,
+			content: pattern.content,
+		};
 	} );
 	const { revertTemplate } = useDispatch( editSiteStore );
 	const isRevertable = isTemplateRevertable( template );
